Move the code of classes, entities and functions to a separate file
