+++ lib/android_store.dart
@@ -1,0 +1,1 @@
+enum AndroidStore { googlePlayStore, apkPure }
